Move gathering column artifacts to separate file
