+++ open_alchemy/models_file/types.py
@@ -24,7 +24,9 @@ class ColumnSchemaArtifacts:
 class ColumnArtifacts:
     """Artifacts for the column portion of a model template."""
 
+    # The name of the column
     name: str
+    # The type of the column
     type: str
 
 
@@ -32,5 +34,22 @@ class ColumnArtifacts:
 class ModelArtifacts:
     """Artifacts for a model template."""
 
+    # The name of the model
     name: str
+    # The columns for the model
     columns: typing.List[ColumnArtifacts]
+
+    # The properties that are required for the TypedDict for the model
+    td_required_props: typing.List[ColumnArtifacts] = dataclasses.field(
+        default_factory=list
+    )
+    # The properties that are not required for the TypedDict for the model
+    td_not_required_propos: typing.List[ColumnArtifacts] = dataclasses.field(
+        default_factory=list
+    )
+    # Whether the required list is empty
+    td_required_empty: bool = False
+    # The name of the TypedDict for required properties
+    td_required_name: typing.Optional[str] = None
+    # The name of the TypedDict for properties that are not required
+    td_not_required_name: typing.Optional[str] = None
