+++ tests/open_alchemy/models_file/model/test_model.py
@@ -7,63 +7,6 @@ from open_alchemy import models_file
 
 
 @pytest.mark.parametrize(
-    "schema, required, expected_artifacts",
-    [
-        (
-            {"type": "type 1"},
-            None,
-            models_file.types.ColumnSchemaArtifacts(type="type 1"),
-        ),
-        (
-            {"type": "type 1", "format": "format 1"},
-            None,
-            models_file.types.ColumnSchemaArtifacts(type="type 1", format="format 1"),
-        ),
-        (
-            {"type": "type 1", "nullable": True},
-            None,
-            models_file.types.ColumnSchemaArtifacts(type="type 1", nullable=True),
-        ),
-        (
-            {"type": "object", "x-de-$ref": "RefModel"},
-            None,
-            models_file.types.ColumnSchemaArtifacts(type="object", de_ref="RefModel"),
-        ),
-        (
-            {"type": "array", "items": {"x-de-$ref": "RefModel"}},
-            None,
-            models_file.types.ColumnSchemaArtifacts(type="array", de_ref="RefModel"),
-        ),
-        (
-            {"type": "type 1"},
-            True,
-            models_file.types.ColumnSchemaArtifacts(type="type 1", required=True),
-        ),
-    ],
-    ids=[
-        "type only",
-        "type with format",
-        "type with nullable",
-        "object",
-        "array",
-        "required given",
-    ],
-)
-@pytest.mark.models_file
-def test_gather_column_artifacts(schema, required, expected_artifacts):
-    """
-    GIVEN schema and required
-    WHEN gather_column_artifacts is called with the schema and required
-    THEN the given expected artifacts are returned.
-    """
-    artifacts = models_file._model.gather_column_artifacts(
-        schema=schema, required=required
-    )
-
-    assert artifacts == expected_artifacts
-
-
-@pytest.mark.parametrize(
     "artifacts, expected_source",
     [
         (
